# serverpackcreator-help/Writerside/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
  /api/v2/modpacks/upload:
    post:
      tags:
      - modpack-controller
      operationId: uploadModPack
      parameters:
      - name: minecraftVersion
        in: query
        required: true
        schema:
          type: string
      - name: modloader
        in: query
        required: true
        schema:
          type: string
      - name: modloaderVersion
        in: query
        required: true
        schema:
          type: string
      - name: startArgs
        in: query
        required: true
        schema:
          type: string
      - name: clientMods
        in: query
        required: true
        schema:
          type: string
      - name: whiteListMods
        in: query
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              required:
              - file
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ZipResponse'
  /api/v2/modpacks/generate:
    post:
      tags:
      - modpack-controller
      operationId: requestGeneration
      parameters:
      - name: modPackID
        in: query
        required: true
        schema:
          type: integer
          format: int32
      - name: minecraftVersion
        in: query
        required: true
        schema:
          type: string
      - name: modloader
        in: query
        required: true
        schema:
          type: string
      - name: modloaderVersion
        in: query
        required: true
        schema:
          type: string
      - name: startArgs
        in: query
        required: true
        schema:
          type: string
      - name: clientMods
        in: query
        required: true
        schema:
          type: string
      - name: whiteListMods
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ZipResponse'
...
  /api/v2/runconfigs/{id}:
    get:
      tags:
      - run-configuration-controller
      operationId: getModpack
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunConfiguration'
...
  /api/v2/modpacks/{id}:
    get:
      tags:
      - modpack-controller
      operationId: getModpack_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModPackView'
  /api/v2/modpacks/download/{id}:
    get:
      tags:
      - modpack-controller
      operationId: downloadModpack
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            application/zip:
              schema:
                type: string
                format: binary
...
components:
  schemas:
    ZipResponse:
      required:
      - message
      - success
      type: object
      properties:
        message:
          type: string
        success:
          type: boolean
        modPackId:
          type: integer
          format: int32
        runConfigId:
          type: integer
          format: int32
        serverPackId:
          type: integer
          format: int32
        status:
          type: string
          enum:
          - QUEUED
          - CHECKING
          - CHECKED
          - GENERATING
          - GENERATED
          - ERROR
...
    ClientMod:
      required:
      - id
      - mod
      type: object
      properties:
        mod:
          type: string
        id:
          type: integer
          format: int32
    RunConfiguration:
      required:
      - clientMods
      - id
      - minecraftVersion
      - modloader
      - modloaderVersion
      - startArgs
      - whitelistedMods
      type: object
      properties:
        id:
          type: integer
          format: int32
        minecraftVersion:
          type: string
        modloader:
          type: string
        modloaderVersion:
          type: string
        startArgs:
          type: array
          items:
            $ref: '#/components/schemas/StartArgument'
        clientMods:
          type: array
          items:
            $ref: '#/components/schemas/ClientMod'
        whitelistedMods:
          type: array
          items:
            $ref: '#/components/schemas/WhitelistedMod'
...
    StartArgument:
      required:
      - argument
      - id
      type: object
      properties:
        argument:
          type: string
        id:
          type: integer
          format: int32
    WhitelistedMod:
      required:
      - id
      - mod
      type: object
      properties:
        mod:
          type: string
        id:
          type: integer
          format: int32
    ModPackView:
      required:
      - dateCreated
      - id
      - name
      - projectID
      - serverPacks
      - sha256
      - size
      - source
      - status
      - versionID
      type: object
      properties:
        name:
          type: string
        id:
          type: integer
          format: int32
        size:
          type: number
          format: double
        source:
          type: string
          enum:
          - ZIP
          - MODRINTH
        status:
          type: string
          enum:
          - QUEUED
          - CHECKING
          - CHECKED
          - GENERATING
          - GENERATED
          - ERROR
        sha256:
          type: string
        serverPacks:
          type: array
          items:
            $ref: '#/components/schemas/ServerPack'
        projectID:
          type: string
        versionID:
          type: string
        dateCreated:
          type: string
          format: date-time
    ServerPack:
      required:
      - confirmedWorking
      - downloads
      - id
      - size
      type: object
      properties:
        id:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        downloads:
          type: integer
          format: int32
        confirmedWorking:
          type: integer
          format: int32
        dateCreated:
          type: string
          format: date-time
        fileID:
          type: integer
          format: int64
        sha256:
          type: string
        runConfiguration:
          $ref: '#/components/schemas/RunConfiguration'
...